Intにパースしてたのを修正

+++ src/main/java/net/akaishi_teacher/mhr/commands/AllTeleportLoc.java
@@ -45,7 +45,7 @@ public class AllTeleportLoc extends MHRAbstractCommand {
 			}
 
 			//Get location.
-			loc = new SimpleLocation(castPlayer(sender).getWorld().getName(), Integer.parseInt(args.get(1)), Integer.parseInt(args.get(2)), Integer.parseInt(args.get(3)), yaw, pitch);
+			loc = new SimpleLocation(castPlayer(sender).getWorld().getName(), Double.parseDouble(args.get(1)), Double.parseDouble(args.get(2)), Double.parseDouble(args.get(3)), yaw, pitch);
 
 			//Teleport horse.
 			for (int i = 0; i < mhr.getStatus().getHorseDatas().size(); i++) {
